pyproject.toml
[project]
name = "assemblyzero-tools"
version = "0.1.0"
description = "AssemblyZero configuration and tooling"
authors = [{name = "Marty McEnroe"}]
readme = "README.md"
license = "PolyForm-Noncommercial-1.0.0"
requires-python = "^3.10"
dependencies = [
    "pywinpty (>=3.0.2,<4.0.0); sys_platform == 'win32'",
    "tzdata (>=2025.3,<2026.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "wexpect (>=4.0.0,<5.0.0); sys_platform == 'win32'",
    "setuptools (>=80.9.0,<81.0.0)",
    "keyring (>=25.7.0,<26.0.0)",
    "anthropic (>=0.76.0,<0.77.0)",
    "langgraph (>=0.2.0,<1.0.0)",
    "langgraph-checkpoint-sqlite (>=1.0.0,<2.0.0)",
    "langchain (>=0.3.0,<1.0.0)",
    "langchain-google-genai (>=2.0.0,<3.0.0)",
    "langchain-anthropic (>=0.3.0,<1.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "google-genai (>=1.60.0,<2.0.0)",
    "pygithub (>=2.8.1,<3.0.0)",
    "tiktoken (>=0.9.0,<1.0.0)"
]

[tool.poetry]
packages = [{include = "assemblyzero"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "mypy (>=1.0.0,<2.0.0)"
]